Emulator: Adds emulator options to serial console
This code allows emulated 'Pong' and 'SysInfo' radio packets to be
sent from the Serial connection by pressin 'q' and 'j'.

This can be used to test a second ground station without having to
change the software.

The 'sendFrame' function in the Radio submodule needed to be made
public for this to be made to work easily.