Emulator: Pass optional data size when sending frame
The existing sendFrame function used 'strlen' to recalculate the
length of the frame data. If the frame data contained a null
value (\0) the frame size would be calculated incorrectly and a
short frame would be transmitted.

Also, added a refactor where the creation and output and sending of
data frame was separated. This will allow the functional control of
frame data in the emulator so that more testing elements can be easily
added. eg. all zero frames, frames with data errors as well as the
existing randon data.